// rinkupay/src/app/models/en.model.ts
/**
 * @autor IIB
 * @date 16/09/2023
 */

/**
  * @description modelo de la coleccion EN
  * 
*/

/**
 * declaracion de la clase como sus variables correspondientes
 */

export class EN {
    _id: string;
    US: string;
    CO: string;
    entregas: number;
    horas: number;
    fecha: Date;
    creacion: Date;
    actualizado: Date;
}